requirements.txt
streamlit
requests
pandas
matplotlib
python-dotenv
openai  # or deepseek if their SDK exists
